Corrected  6-size.c

+++ 0x00-hello_world/6-size.c
@@ -13,7 +13,7 @@ int main(void)
 	printf("Size of Char in bytes: 1%d bytes\n", sizeof(char));
 	printf("Size of int in bytes: 1%d bytes\n", sizeof(int));
 	Printf("Size of long int in bytes: 1%d bytes\n", sizeof(long));
-	printf("Size of long long int in bytes: 1%d bytes\n", sizeof(double));
+	printf("Size of long long int in bytes: 1%d bytes\n", sizeof()long long int);
 	printf("Size of float in bytes: 1%d bytes\n", sizeof(float));
 	return (0);
 }
